# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Section:
    properties:
      id:
        type: integer
      name:
        type: string
      restaurant_id:
        type: integer
    type: object
  models.Table:
    properties:
      id:
        type: integer
      number_of_table:
        type: integer
      qr:
        type: string
      section_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API сервера для системы управления рестораном
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Restaurant Management System API
  version: "1.0"
paths:
...
  /sections:
    post:
      consumes:
      - application/json
      description: Создает новую секцию для указанного ресторана
      parameters:
      - description: Данные секции
        in: body
        name: section
        required: true
        schema:
          $ref: '#/definitions/models.Section'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Создать новую секцию ресторана
      tags:
      - sections
  /sections/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет секцию по её ID
      parameters:
      - description: ID секции
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Удалить секцию
      tags:
      - sections
    get:
      consumes:
      - application/json
      description: Возвращает секцию по её ID
      parameters:
      - description: ID секции
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Section'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Получить секцию по ID
      tags:
      - sections
    put:
      consumes:
      - application/json
      description: Обновляет данные существующей секции
      parameters:
      - description: ID секции
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные секции
        in: body
        name: section
        required: true
        schema:
          $ref: '#/definitions/models.Section'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Обновить данные секции
      tags:
      - sections
  /sections/restaurant/{restaurantID}:
    get:
      consumes:
      - application/json
      description: Возвращает список секций для указанного ресторана
      parameters:
      - description: ID ресторана
        in: path
        name: restaurantID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Section'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Получить секции по ID ресторана
      tags:
      - sections
  /tables:
    post:
      consumes:
      - application/json
      description: Создает новый столик для указанной секции
      parameters:
      - description: Данные столика
        in: body
        name: table
        required: true
        schema:
          $ref: '#/definitions/models.Table'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Создать новый столик
      tags:
      - tables
  /tables/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет столик по его ID
      parameters:
      - description: ID столика
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Удалить столик
      tags:
      - tables
    get:
      consumes:
      - application/json
      description: Возвращает столик по его ID
      parameters:
      - description: ID столика
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Table'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Получить столик по ID
      tags:
      - tables
    put:
      consumes:
      - application/json
      description: Обновляет данные существующего столика
      parameters:
      - description: ID столика
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные столика
        in: body
        name: table
        required: true
        schema:
          $ref: '#/definitions/models.Table'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Обновить данные столика
      tags:
      - tables
  /tables/{id}/qr:
    post:
      consumes:
      - application/json
      description: Генерирует и возвращает QR-код для указанного столика
      parameters:
      - description: ID столика
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Сгенерировать QR-код для столика
      tags:
      - tables
  /tables/section/{sectionID}:
    get:
      consumes:
      - application/json
      description: Возвращает список столиков для указанной секции
      parameters:
      - description: ID секции
        in: path
        name: sectionID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Table'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Получить столики по ID секции
      tags:
      - tables
...
swagger: "2.0"